save diary test ok

+++ src/presentation/mutate/request.rs
@@ -3,9 +3,9 @@ use serde::Deserialize;
 #[derive(Deserialize)]
 pub struct MutateRequest {
     #[serde(rename = "clientId")]
-    pub client_id: usize,
+    pub client_id: i32,
     #[serde(rename = "targetText")]
     pub target_text: Vec<String>,
     #[serde(rename = "mutatedLength")]
-    pub mutated_length: Option<usize>,
+    pub mutated_length: Option<i32>,
 }
